Add _form partial for create and edit template

+++ src/Controller/PinsController.php
@@ -84,11 +84,14 @@ class PinsController extends AbstractController
      * @Route("/pins/{id<[0-9]+>}/delete", name="app_pins_delete", methods="DELETE")
      */
 
-    public function delete(EntityManagerInterface $em, Pin $pin): Response
+    public function delete(Request $request, EntityManagerInterface $em, Pin $pin): Response
     {
+        
+        if ($this->isCsrfTokenValid('pin_deletion_'. $pin->getId(), $request->request->get('csrf_token') )) {
         $em->remove($pin);
         $em->flush();
-
+        
+        } 
         return $this->redirectToRoute('app_home');
     }
